backend/requirements.txt
Django>=5.0.6
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
redis>=5.0.1
django-redis==5.3.0
python-dotenv>=1.0.1
django-cors-headers>=4.3.1
django-user-agents==0.4.0
gunicorn>=21.2.0
